Updating windows bootstrap
Prior to this commit choco and wmf were not installed globally on all
windows nodes, automatically.  This led to the first time 'baseline' was
applied the windows nodes would reboot due to wmf installation.  This
commit makes this happen the first agent run of puppet.  Also, the
windows firewall now gets configured to allow ICMP, RDP, and WinRM.