Clean up profile::puppet::master::hiera
  * Clean up version deliniation code
  * Switch to
      * Minimal global hiera.yaml
      * Environment hiera.yaml
  * Remove exteraneous, old configuration not applicable to demo env

We want to use an environment-level hiera.yaml because this is a best
practice with Hiera 5. Part of using a hiera.yaml directly in the
control-repo is minimalizing the global configuration. It made sense
when making those changes to clean up all the things we A) aren't using,
or B) are old, arguably bad ways to deal with hiera, e.g.
synchronization of keys between masters using Puppet's fileserver.conf
functionality instead of file-sync.