Refactor role::master_server > role::seteam_master
I have a need to introduce a new role for a more general, standard
Puppet master role, one that can be configured and integrated with
standard supporting services such as Gitlab, Github, CD4PE, and so
forth.

The previously existing role::master_server did not meet these
requirements, and so I needed to create a new role, which I am calling
role::puppet_master.

Once role::puppet_master exists though, it can be confused with
role::master_server. In order to reduce confusion in naming I am
creating a new role to replace role::master_server. The replacement is
role::seteam_puppet_master.

The "seteam" prefix conveys that this is a variant of a "normal" puppet
master. Using "seteam" specifically communicates the business purpose of
it being a variant for and used by the seteam.

The old role, role::master_server, is retained in this commit for
temporary backwards compatility and has been modified to simply contain
the new role, role::seteam_puppet_master.

Next steps will be to modify the build system to use
role::seteam_puppet_master directly, so that the deprecated
role::master_server can be removed from the control repository.