Add role::require_kernel function
This allows roles to cleanly assert a required kernel version even if
the profiles or other content included does not happen to include that
check. This is necessary to have roles that pass simple platform tests
when there isn't an obvious place to specify the dependency outside of
the role itself.